requirements.txt
Flask>=2.0
python-dotenv>=0.21.0
requests>=2.28.0
whois>=0.9.20190521
cachetools>=5.0
